requirements.txt
flask==2.3.3
werkzeug>=2.3.7
python-dotenv==1.0.0
langchain-groq==0.2.3
langchain-core==0.3.30
groq==0.15.0
requests==2.31.0
flask-cors==3.0.10
